+++ record.h
@@ -1,5 +1,5 @@
 #ifndef record_h_INCLUDED
 #define record_h_INCLUDED
 
-void record(long double in_size,float entropy,long double out_size,double time);
+void record(long int in_size,float entropy,long int out_size,double time);
 #endif //record_h_INCLUDED
